Refactor API
Improve species class (DRY)

Clean up module structure

Split datasets' docstrings and run functions
