+++ atomdb/test/test_numeric.py
@@ -22,14 +22,18 @@
 #
 # --
 
+from importlib.resources import files
+
+import os
 
 import pytest
 
-from numpy.testing import assert_equal, assert_almost_equal
-from importlib_resources import files
 import numpy as np
-from atomdb.api import load
-import os
+
+from numpy.testing import assert_almost_equal
+
+from atomdb import load
+
 
 # get test data path
 TEST_DATAPATH = files("atomdb.test.data")
